Refactor CallConvKind to strongly typed enum class
Replaced `CallConvKind` with a strongly-typed `enum class` for improved type safety and clarity. Updated references to use explicit casts where necessary and adjusted global `g_CallConv` initialization to match the new structure. This change ensures stricter type checking and reduces potential errors.

+++ src/Semantic/Type/TypeInfoSpecifics.cpp
@@ -1011,7 +1011,7 @@ uint32_t TypeInfoFuncAttr::numTotalRegisters() const
 
 const CallConv& TypeInfoFuncAttr::getCallConv() const
 {
-    return g_CallConv[callConv];
+    return g_CallConv[static_cast<int>(callConv)];
 }
 
 TypeInfoParam* TypeInfoStruct::findChildByNameNoLock(const Utf8& childName) const
